+++ src/main/java/com/idt/aio/repository/SimilarityDocRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SimilarityDocRepository extends JpaRepository<SimilarityDoc, Long> {
+public interface SimilarityDocRepository extends JpaRepository<SimilarityDoc, Integer> {
+
 }
